Extract team access methods to a helper.
This change only removes duplication, the functionality is still the same.
